Raw mode redux.
It turns out that Termion wraps stuff around the escape codes. Since I found that to be the most fun part of Bernhardt's screencast. Unfortunately, Rust doesn't have some of the facilities of Ruby when setting raw mode in the terminal or even getting a single character from stdin. I had to go lower level than the screencast for those things; reading from stdin only supports ASCII because I only know how to read a single byte. It could be that it's easy enough to detect the particular bytes for UTF-8. 😬

I really should have documented my exploration of termios. One thing that confused me is that I thought I had to set the modes on both the stdin and stdout file descriptors but it turns out that they refer to the ptty -- so they're the same! Doing it twice really messed me up. I spent hours trying to figure out why I couldn't reset the terminal but it was because I was doing everything twice that I was resetting then reresetting and thereby doing nothing.