18:22 Navigate the buffer.
I'm able to use keys to move around the buffer. I didn't use the same keys as Bernhardt's screencast because Ctrl-b interferes with tmux. Instead I used the same keys as vim to go up, down, left, and right but they're modified with Ctrl.

I added a "key inspector" command to the program. Since I had to do lower level stuff with the raw terminal I need to know the byte value of each key. I'm kind of bummed about it, but I did look into parsing UTF-8 and it may not be *that* difficult. The bytes are obviously well defined; I'm just not used to doing stuff like that.

https://en.wikipedia.org/wiki/UTF-8#Description