Done. Implemented undo.
I tried to make this less repetitive but I kept running into ownership problems. I spent an hour trying to suss it out but forget about it.

The code tracks pretty closely to the screencast but it's not idiomatic Rust (not that I could even write idiomatic Rust at this point) so I think there's a lot that could be improved. If I keep working on this to try and write a usable editor, then I would probably start over.